feature(13-dispatch-custom-events): use  createEventDispatcher()
So we guess, we kind of need to handle the submit-event of the form
inside the parent-component, not necessarily in form-component itself.

Now, you maybe might thinking we could use event-forwarding on the
'on:submit' event, like we saw inside 'Modal.svelte', we use
event-forwarding on the 'on:click' and we handle in the
parent-component.

The problem with that, is not allow us to send the data that we have from
the form up the component tree (parent-component). Because when we use
event-forwarding we don't pass any data, we literally forward the event
whatever what that is. So we couldn't add people to the array with this
event-forwarding.

Instead we're going to need to **dispatch**, what's known as
a custom-event from our form; and then send with it the form data to the
parent component (App.svelte).