feature(ninjapolls): casting vote
Since the PollDetails-component is nested component, we create
inline-function, because we need to pass data onto another function. We
need to **dispatch** the custom event with that event to App.svelte. We
use createEventDispatcher to emit that event.

We can't directly access that event on PollDetails-component, because
PollDetails-component not nested directly in App.svelte. It's nested
inside the PollList-component.

We have to listen PollDetails-event in PollList-component. We don't
really want to react the event on the PollList-component. We don't want
to edit the polls, we want to edit the actual data on the App.svelte. We
need to then forward it to the parent-component which is App.svelte.

We've seen how we can do **event-forwarding** with click-event and
submit-event. We can do the same with **custom-event**. So we just
assign a handler with "on:vote".

Therefore, in App.svelte we can listen to that event on the <PollList/>,
because it's forwarded. We create a function called "handleVote" to
handle event-forwarding on App.svelte.

Schema:

PollDetails.svelte --> PollList.svelte --> App.svelte